+++ versions/1.8.9/src/main/java/io/github/axolotlclient/mixin/ClientPlayerEntityMixin.java
@@ -22,13 +22,14 @@
 
 package io.github.axolotlclient.mixin;
 
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import io.github.axolotlclient.modules.hud.HudManager;
 import io.github.axolotlclient.modules.hud.gui.hud.simple.ToggleSprintHud;
 import net.minecraft.client.entity.living.player.LocalClientPlayerEntity;
 import net.minecraft.client.options.KeyBinding;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Redirect;
 
 @Mixin(LocalClientPlayerEntity.class)
 public abstract class ClientPlayerEntityMixin {
@@ -39,9 +40,9 @@ public abstract class ClientPlayerEntityMixin {
 	 * @author DragonEggBedrockBreaking
 	 * <p>License: MPL-2.0</p>
 	 */
-	@Redirect(method = "mobTick", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/options/KeyBinding;isPressed()Z"))
-	private boolean axolotlclient$alwaysPressed(KeyBinding sprintKey) {
+	@WrapOperation(method = "mobTick", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/options/KeyBinding;isPressed()Z"))
+	private boolean axolotlclient$alwaysPressed(KeyBinding sprintKey, Operation<Boolean> original) {
 		ToggleSprintHud hud = (ToggleSprintHud) HudManager.getInstance().get(ToggleSprintHud.ID);
-		return hud.getSprintToggled().get() || sprintKey.isPressed();
+		return hud.getSprintToggled().get() || original.call(sprintKey);
 	}
 }